Adicionado testes para a página de login de usuários.
