+++ mercado/base/templates/base/perfil.html
@@ -5,10 +5,12 @@
     <div class="col-3"></div>
     <div class="col">
         <h1> Perfil do Usuário</h1>
+        {% if user.is_authenticated %}
         <form method="post" action="{% url 'logout' %}">
             {% csrf_token %}
             <button type="submit" class="btn btn-danger col-12">Sair</button>
         </form>
+        {% endif %}
     </div>
     <div class="col-3"></div>
 </div>
